Add user associations with articles and build signed_in, signed_out, owner's different cases

+++ app/controllers/articles_controller.rb
@@ -21,7 +21,12 @@ class ArticlesController < ApplicationController
     end
   end
 
+  def edit
+    redirect_to root_path, alert: 'You can only edit your own articles.' unless @article.user == current_user
+  end
+
   def update
+    redirect_to root_path, alert: 'You can only edit your own articles.' unless @article.user == current_user
     if @article.update(article_params)
       redirect_to article_path(@article.id), notice: 'Article has been edited'
     else
@@ -31,6 +36,7 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path, alert: 'You can only delete your own articles.' unless @article.user == current_user
     redirect_to articles_path, notice: 'Article has been deleted' if @article.destroy
   end
